chore(studio): remove deprecated Query Insights / supamonitor feature
supamonitor is fully deprecated and no real project should have it
enabled anymore, so removes the entire Query Insights page and its
BigQuery-backed supamonitor_logs report (28 files, ~3.5k lines).

Query Performance (pg_stat_statements-based) becomes the only nav
item, replacing the old conditional between it and Query Insights.
Two generic SQL-parsing helpers (getTableName/getColumnName) and the
ClassifiedQuery type were relocated into QueryPerformance since its
own AI-explanation prompt builder depended on them.

If a project still has the supamonitor extension loaded (checked live
via shared_preload_libraries, e.g. some internal project), the Query
Performance page now shows a non-blocking warning instead of routing
to the removed feature or crashing.

Added a permanent redirect from the old /query-insights URL to
/query-performance for any bookmarked links.
