+++ apps/studio/components/interfaces/Observability/Observability.constants.ts
@@ -3,7 +3,6 @@ import { DOCS_URL } from '@/lib/constants'
 export const OBSERVABILITY_DOCS_HREFS = {
   overview: `${DOCS_URL}/guides/telemetry/reports`,
   queryPerformance: `${DOCS_URL}/guides/platform/performance#examining-query-performance`,
-  queryInsights: `${DOCS_URL}/guides/platform/performance#examining-query-performance`,
   apiGateway: `${DOCS_URL}/guides/telemetry/reports#api-gateway`,
   database: `${DOCS_URL}/guides/telemetry/reports#database`,
   dataApi: `${DOCS_URL}/guides/telemetry/reports#postgrest`,
