--- apps/studio/components/interfaces/QueryInsights/QueryInsightsHealth/QueryInsightsHealthScoreSkeleton.tsx
@@ -1,9 +1,0 @@
-export const QueryInsightsHealthScoreSkeleton = () => (
-  <>
-    <div className="h-12 w-12 rounded-full bg-surface-300 animate-pulse" />
-    <div className="flex flex-col gap-1.5">
-      <div className="h-3 w-20 rounded-sm bg-surface-300 animate-pulse" />
-      <div className="h-5 w-24 rounded-sm bg-surface-300 animate-pulse" />
-    </div>
-  </>
-)
