--- apps/studio/components/interfaces/QueryInsights/QueryInsightsTable/QueryInsightsTable.types.ts
@@ -1,2 +1,0 @@
-export type Mode = 'triage' | 'explorer'
-export type IssueFilter = 'all' | 'error' | 'index' | 'slow'
