+++ apps/studio/components/interfaces/QueryPerformance/QueryPerformance.ai.ts
@@ -1,10 +1,6 @@
-import { QueryPerformanceRow } from './QueryPerformance.types'
+import { ClassifiedQuery, QueryPerformanceRow } from './QueryPerformance.types'
+import { getColumnName, getTableName } from './QueryPerformance.utils'
 import type { QueryPlanRow } from '@/components/interfaces/ExplainVisualizer/ExplainVisualizer.types'
-import type { ClassifiedQuery } from '@/components/interfaces/QueryInsights/QueryInsightsHealth/QueryInsightsHealth.types'
-import {
-  getColumnName,
-  getTableName,
-} from '@/components/interfaces/QueryInsights/QueryInsightsTable/QueryInsightsTable.utils'
 
 export interface QueryExplanationPrompt {
   query: string
